Add drawer navigation and split inventory views
