+++ frontend/src/components/pos/TopBar.tsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button } from '../ui/button';
 import { Badge } from '../ui/badge';
@@ -12,6 +13,8 @@ interface TopBarProps {
   onNavigateAnalytics?: () => void;
   onNavigatePos?: () => void;
   isAnalytics?: boolean;
+  onNavigateProducts?: () => void;
+  isProducts?: boolean;
   onNavigateInventory?: () => void;
   isInventory?: boolean;
   onNavigatePurchases?: () => void;
@@ -28,6 +31,8 @@ export function TopBar({
   onNavigateAnalytics,
   onNavigatePos,
   isAnalytics,
+  onNavigateProducts,
+  isProducts,
   onNavigateInventory,
   isInventory,
   onNavigatePurchases,
@@ -40,11 +45,64 @@ export function TopBar({
   const { t, i18n } = useTranslation();
   const { theme, setTheme } = useTheme();
   const displayName = useAuthStore((state) => state.displayName);
-  const role = useAuthStore((state) => state.role);
   const storeName = useStoreProfileStore((state) => state.name);
   useStoreProfileQuery();
 
-  const canManageInventory = role?.toLowerCase() === 'admin' || role?.toLowerCase() === 'manager';
+  const [isDrawerOpen, setIsDrawerOpen] = useState(false);
+
+  const isPosActive =
+    Boolean(onNavigatePos) &&
+    !isAnalytics &&
+    !isInventory &&
+    !isPurchases &&
+    !isInvoices &&
+    !isSettings &&
+    !isProducts;
+
+  const navItems = [
+    {
+      key: 'pos',
+      label: t('pos'),
+      onClick: onNavigatePos,
+      isActive: isPosActive
+    },
+    {
+      key: 'analytics',
+      label: t('analytics'),
+      onClick: onNavigateAnalytics,
+      isActive: Boolean(isAnalytics)
+    },
+    {
+      key: 'products',
+      label: t('products'),
+      onClick: onNavigateProducts,
+      isActive: Boolean(isProducts)
+    },
+    {
+      key: 'inventory',
+      label: t('inventory'),
+      onClick: onNavigateInventory,
+      isActive: Boolean(isInventory)
+    },
+    {
+      key: 'invoices',
+      label: t('invoices'),
+      onClick: onNavigateInvoices,
+      isActive: Boolean(isInvoices)
+    },
+    {
+      key: 'purchases',
+      label: t('purchases'),
+      onClick: onNavigatePurchases,
+      isActive: Boolean(isPurchases)
+    },
+    {
+      key: 'settings',
+      label: t('settings'),
+      onClick: onNavigateSettings,
+      isActive: Boolean(isSettings)
+    }
+  ].filter((item) => Boolean(item.onClick));
 
   const toggleLanguage = () => {
     const next = i18n.language === 'en' ? 'ar' : 'en';
@@ -55,66 +113,39 @@ export function TopBar({
     setTheme(theme === 'dark' ? 'light' : 'dark');
   };
 
+  const openDrawer = () => setIsDrawerOpen(true);
+  const closeDrawer = () => setIsDrawerOpen(false);
+  const handleNavigate = (callback?: () => void) => {
+    if (!callback) {
+      return;
+    }
+    callback();
+    closeDrawer();
+  };
+
   return (
     <div className="flex flex-wrap items-center justify-between gap-3 rounded-xl bg-white p-4 shadow-sm dark:bg-slate-900">
-      <div>
-        <h1 className="text-2xl font-bold text-emerald-600 dark:text-emerald-300">{storeName}</h1>
-        <p className="text-xs text-slate-500">Asia/Beirut</p>
-        {lastScan && (
-          <Badge className="mt-2">Last scan: {lastScan}</Badge>
+      <div className="flex items-start gap-3">
+        {navItems.length > 0 && (
+          <Button
+            type="button"
+            className="bg-slate-200 text-slate-900 hover:bg-slate-300 dark:bg-slate-700 dark:text-slate-100"
+            onClick={openDrawer}
+            aria-expanded={isDrawerOpen}
+            aria-controls="topbar-drawer"
+          >
+            {t('menu')}
+          </Button>
         )}
+        <div>
+          <h1 className="text-2xl font-bold text-emerald-600 dark:text-emerald-300">{storeName}</h1>
+          <p className="text-xs text-slate-500">Asia/Beirut</p>
+          {lastScan && (
+            <Badge className="mt-2">Last scan: {lastScan}</Badge>
+          )}
+        </div>
       </div>
       <div className="flex flex-wrap items-center gap-2">
-        {onNavigateAnalytics && !isAnalytics && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigateAnalytics}>
-            {t('analytics')}
-          </Button>
-        )}
-        {canManageInventory && onNavigateInvoices && !isInvoices && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigateInvoices}>
-            {t('invoices')}
-          </Button>
-        )}
-        {canManageInventory && onNavigatePurchases && !isPurchases && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigatePurchases}>
-            {t('purchases')}
-          </Button>
-        )}
-        {canManageInventory && onNavigateInventory && !isInventory && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigateInventory}>
-            {t('products')}
-          </Button>
-        )}
-        {canManageInventory && onNavigateSettings && !isSettings && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigateSettings}>
-            {t('settings')}
-          </Button>
-        )}
-        {onNavigatePos && isAnalytics && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigatePos}>
-            {t('backToPos')}
-          </Button>
-        )}
-        {onNavigatePos && isInvoices && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigatePos}>
-            {t('backToPos')}
-          </Button>
-        )}
-        {onNavigatePos && isPurchases && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigatePos}>
-            {t('backToPos')}
-          </Button>
-        )}
-        {onNavigatePos && isInventory && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigatePos}>
-            {t('backToPos')}
-          </Button>
-        )}
-        {onNavigatePos && isSettings && (
-          <Button type="button" className="bg-emerald-500 hover:bg-emerald-400" onClick={onNavigatePos}>
-            {t('backToPos')}
-          </Button>
-        )}
         <Button type="button" className="bg-slate-200 text-slate-900 hover:bg-slate-300 dark:bg-slate-700 dark:text-slate-100" onClick={toggleTheme}>
           {theme === 'dark' ? t('lightMode') : t('darkMode')}
         </Button>
@@ -126,6 +157,51 @@ export function TopBar({
           {t('logout')}
         </Button>
       </div>
+      {isDrawerOpen && (
+        <div className="fixed inset-0 z-50" role="dialog" aria-modal="true">
+          <div
+            className="absolute inset-0 bg-slate-900/40"
+            onClick={closeDrawer}
+            aria-hidden="true"
+          />
+          <aside
+            id="topbar-drawer"
+            className="absolute left-0 top-0 flex h-full w-72 flex-col gap-6 bg-white p-6 shadow-xl dark:bg-slate-900"
+          >
+            <div className="flex items-center justify-between">
+              <h2 className="text-lg font-semibold text-slate-900 dark:text-slate-100">{t('menu')}</h2>
+              <Button
+                type="button"
+                className="bg-slate-200 text-slate-900 hover:bg-slate-300 dark:bg-slate-700 dark:text-slate-100"
+                onClick={closeDrawer}
+              >
+                ×
+              </Button>
+            </div>
+            <nav className="flex flex-1 flex-col gap-2">
+              {navItems.length === 0 ? (
+                <p className="text-sm text-slate-500 dark:text-slate-400">{t('menuEmpty')}</p>
+              ) : (
+                navItems.map((item) => (
+                  <button
+                    key={item.key}
+                    type="button"
+                    onClick={() => handleNavigate(item.onClick)}
+                    className={`rounded-md px-4 py-2 text-left text-sm font-medium transition ${
+                      item.isActive
+                        ? 'bg-emerald-500 text-white hover:bg-emerald-400'
+                        : 'bg-slate-100 text-slate-700 hover:bg-slate-200 dark:bg-slate-800 dark:text-slate-100 dark:hover:bg-slate-700'
+                    }`}
+                    aria-current={item.isActive ? 'page' : undefined}
+                  >
+                    {item.label}
+                  </button>
+                ))
+              )}
+            </nav>
+          </aside>
+        </div>
+      )}
     </div>
   );
 }
