+++ frontend/src/i18n/index.ts
@@ -9,6 +9,9 @@ const resources = {
       username: 'Username',
       password: 'Password',
       logout: 'Logout',
+      menu: 'Menu',
+      menuEmpty: 'No navigation options available yet.',
+      pos: 'Point of Sale',
       retry: 'Retry',
       searchProducts: 'Search products or scan barcode',
       pinnedBadge: 'Pinned',
@@ -32,6 +35,7 @@ const resources = {
       purchases: 'Purchases',
       invoices: 'Invoices',
       products: 'Products',
+      inventory: 'Inventory',
       backToPos: 'Back to POS',
       profitLeaders: 'Top profit SKUs',
       lossLeaders: 'Loss leaders',
@@ -106,6 +110,9 @@ const resources = {
       inventorySummaryItemsDescription: 'Totals for each catalog item.',
       inventorySummaryItemsEmpty: 'No item totals available yet.',
       inventorySummaryQuantityLabel: '{{count}} on hand',
+      inventoryViewModeLabel: 'Show stock value by',
+      inventoryViewModeOptionCategories: 'Category',
+      inventoryViewModeOptionItems: 'Item',
       inventoryFormIncomplete: 'Please fill in all required fields.',
       inventoryCreateSuccess: 'Product created successfully.',
       inventoryCreateError: 'Unable to create product.',
@@ -190,6 +197,9 @@ const resources = {
       username: 'اسم المستخدم',
       password: 'كلمة المرور',
       logout: 'تسجيل الخروج',
+      menu: 'القائمة',
+      menuEmpty: 'لا توجد خيارات تنقّل متاحة حالياً.',
+      pos: 'نقطة البيع',
       retry: 'أعد المحاولة',
       searchProducts: 'ابحث عن المنتجات أو امسح الباركود',
       pinnedBadge: 'مُثبَّت',
@@ -230,6 +240,7 @@ const resources = {
       analyticsError: 'تعذر تحميل التحليلات. يتم عرض بيانات تجريبية.',
       forecastFor: 'توقع لـ',
       products: 'المنتجات',
+      inventory: 'الجرد',
       inventoryAddTitle: 'إضافة منتج',
       inventoryAddSubtitle: 'أدخل تفاصيل المنتج لتحديث الكتالوج.',
       inventoryName: 'اسم المنتج',
@@ -287,6 +298,9 @@ const resources = {
       inventorySummaryItemsDescription: 'إجماليات كل صنف في الكتالوج.',
       inventorySummaryItemsEmpty: 'لا توجد بيانات للأصناف بعد.',
       inventorySummaryQuantityLabel: '{{count}} في المخزون',
+      inventoryViewModeLabel: 'عرض قيمة المخزون بحسب',
+      inventoryViewModeOptionCategories: 'الفئة',
+      inventoryViewModeOptionItems: 'العنصر',
       inventoryFormIncomplete: 'يرجى تعبئة جميع الحقول المطلوبة.',
       inventoryCreateSuccess: 'تم إنشاء المنتج بنجاح.',
       inventoryCreateError: 'تعذر إنشاء المنتج.',
