+++ frontend/src/pages/POSPage.tsx
@@ -618,6 +618,7 @@ export function POSPage() {
         onNavigateAnalytics={canSeeAnalytics ? () => navigate('/analytics') : undefined}
         onNavigateInvoices={canManageInventory ? () => navigate('/invoices') : undefined}
         onNavigatePurchases={canManageInventory ? () => navigate('/purchases') : undefined}
+        onNavigateProducts={canManageInventory ? () => navigate('/products') : undefined}
         onNavigateInventory={canManageInventory ? () => navigate('/inventory') : undefined}
         onNavigateSettings={canManageInventory ? () => navigate('/settings') : undefined}
       />
